requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
bcrypt
pydantic
python-multipart
python-dotenv
pandas
openpyxl
email-validator
pydantic[email]
passlib[bcrypt]
python-jose
passlib
bcrypt
googlesearch-python
serpapi